vo_gl: add "backend" suboption to select x11/sdl etc backend
Add new "backend" suboption that allows selecting the display
backend used by vo_gl. Normally, it's auto-selected, but sometimes
it's desireable to explicitly select it.

Remove the gl_sdl VO. SDL backend can now be selected with
--vo=gl:backend=sdl

This is based on svn commit 34438, and tries to be compatible with it.
The undocumented numeric backend names serve this purpose. (They are
undocumented because names are preferred.)