demux: do not set global sub_utf8
demux_lavf and demux_mkv used to set the global variable sub_utf8 to
signal the character set of the subtitles they output. The main
purpose of this variable is to hold the value of the --utf8 option
which affects how external subtitle files are interpreted. Now that
subtitle rendering is done through libass only, this variable no
longer has any effect outside the parsing of those external files.
Remove the code setting the variable from the demuxers.